myadi

+++ app/Database/Migrations/2024-05-09-160901_CreateItems.php
@@ -22,11 +22,11 @@ class CreateItems extends Migration
                 'unique' => true
             ],
             'quantity' => [
-                'type' => 'DOUBLE',
+                'type' => 'FLOAT',
                 'null' => true,
             ],
             'price_per_unit' => [
-                'type' => 'DOUBLE',
+                'type' => 'FLOAT',
                 'null' => true,
             ],
         ]);
