+++ app/Database/Migrations/2024-05-09-161216_CreateSales.php
@@ -27,11 +27,11 @@ class CreateSales extends Migration
                 'null' => true,
             ],
             'quantity' => [
-                'type' => 'DOUBLE',
+                'type' => 'FLOAT',
                 'null' => true,
             ],
             'price_per_unit' => [
-                'type' => 'DOUBLE',
+                'type' => 'FLOAT',
                 'null' => true,
             ],
         ]);
